# docs/swagger.yaml
components:
  schemas:
...
    dto.CreateTaskRequest:
      properties:
        assignee_id:
          type: string
        column_id:
          type: string
        description:
          type: string
        due_date:
          type: string
        priority:
          type: string
        title:
          type: string
      type: object
...
    dto.TaskAssigneeResponse:
      properties:
        avatar_key:
          type: string
        display_name:
          type: string
        id:
          type: string
        login:
          type: string
      type: object
    dto.TaskResponse:
      properties:
        assignee:
          $ref: '#/components/schemas/dto.TaskAssigneeResponse'
        column_id:
          type: string
        created_at:
          type: string
        description:
          type: string
        due_date:
          type: string
        id:
          type: string
        position:
          type: integer
        priority:
          type: string
        title:
          type: string
        updated_at:
          type: string
      type: object
...
    dto.UpdateTaskRequest:
      properties:
        assignee_id:
          type: string
        column_id:
          type: string
        description:
          type: string
        due_date:
          type: string
        position:
          type: integer
        priority:
          type: string
        title:
          type: string
      type: object
...
info:
  description: Collaborative task management API for development teams.
  title: Tandem API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/workspaces/{id}/boards/{boardId}/tasks:
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateTaskRequest'
                description: Task payload
                summary: request
        description: Task payload
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.TaskResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Create a task
      tags:
      - tasks
  /api/v1/workspaces/{id}/boards/{boardId}/tasks/{taskId}:
    delete:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    patch:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateTaskRequest'
                description: Task payload
                summary: request
        description: Task payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.TaskResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
...
servers:
- url: http://localhost:8080/